requirements.txt
aiobotocore==2.3.4
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
alembic==1.8.1
apispec==5.2.2
async-timeout==4.0.2
attrs==22.1.0
boto3==1.24.42
botocore==1.24.21
cachetools==5.0.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
commonmark==0.9.1
cramjam==2.5.0
cryptography==37.0.4
databricks-cli==0.17.0
deepdiff==5.8.1
dill==0.3.2
distlib==0.3.5
dnslib==0.9.20
dnspython==2.2.1
docker==5.0.3
ecdsa==0.18.0
entrypoints==0.4
fastparquet==0.8.1
filelock==3.7.1
Flask==2.1.3
frozenlist==1.3.0
fsspec==2022.7.1
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
localstack==1.0.3
localstack-client==1.35
localstack-ext==1.0.3
Mako==1.2.1
MarkupSafe==2.1.1
mlflow==1.27.0
multidict==6.0.2
numpy==1.23.1
oauthlib==3.2.0
ordered-set==4.1.0
packaging==21.3
pandas==1.4.3
pbr==5.9.0
pipenv==2022.7.24
platformdirs==2.5.2
pluggy==1.0.0
plux==1.3.1
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.2
protobuf==4.21.3
psutil==5.9.1
py==1.11.0
pyaes==1.6.1
pyarrow==8.0.0
pyasn1==0.4.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.1
PyYAML==6.0
querystring-parser==1.2.4
requests==2.25.1
rich==12.5.1
rsa==4.9
s3fs==2022.7.1
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.9.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
stevedore==4.0.0
tabulate==0.8.10
tailer==0.4.1
threadpoolctl==3.1.0
tomli==2.0.1
urllib3==1.26.11
virtualenv==20.16.2
virtualenv-clone==0.5.7
websocket-client==1.3.3
Werkzeug==2.2.0
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.1
